Add files via upload
Saturated Reverse Polish Notation (SRPN calculator). This was a project I completed in the one University module that I completed. The assignment presented a calculator which we could interact with, but we could not see the underlying code. The calculator contained many bugs and flaws and my task was to replicate it in Java as best as possible. At this stage in the course we had not yet been taught object orientated principles so I completed the task in a procedural manner. I scored the equivalent of an 'A' grade for this.