Add files via upload
This is a game of connect four which you can play against the computer. The computer just takes a go by randomly selecting a column. The code has been written so that the game could be easily scaled to include a larger board size and more players. This was a learning exercise in the concept of modularity, with my aim being to break the game down into various self contained elements which could be replaced with more sophisticated aspects in the future (for instance the display module could be changed to a graphics engine instead of simply printing out to the console).