Add files via upload
This is a two player game of tic tac toe that I built in MIPS. It stores the board, asks for user input, updates the board, checks for a winner or a draw, and then asks for the next player to take their go (assuming the game hasn't ended). It's a pretty simple game but this was a great exercise for bringing together everything I'd learned about MIPS into a slightly larger project.